pyproject.toml
[tool.poetry]
name = "vocalize-bot"
version = "0.1.0"
description = "A discord self-powered tts bot"
authors = ["Your Name <you@example.com>"]
license = "MIT license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = {extras = ["voice"], version = "^2.3.2"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
